refactoring Modal code, moving Toast to page.tsx
